+++ app/Http/Requests/ImportTrackingProviderLocationsRequest.php
@@ -14,6 +14,8 @@ class ImportTrackingProviderLocationsRequest extends BaseRequest
         return [
             'merchant_id' => ['required', 'uuid', 'exists:merchants,uuid'],
             'only_with_geofences' => ['sometimes', 'boolean'],
+            'locations' => ['nullable', 'array', 'min:1'],
+            'locations.*.provider_location_id' => ['required_with:locations', 'string'],
         ];
     }
 }